No more shortcodes!
absimg, absvideo, and abslink no longer function as "make an absolute path" shortcuts; instead, the content file is evaluated as a template, so it can simply include a reference to the site's BaseURL.

(This does introduce an annoying quirk in the content itself: the template delimiter '%' needs to be escaped in normal usage.  Oh, well.)

And by adding some HTML attribute and media-handling extensions, Markdig now understands how to render appropriate in-line img and video tags.