Progress checkpoint: `SiteDataReference` ... * Takes the place of string-based references in PostData, so that front matter references can be resolved later. * Implements a pre-resolved state, prior to looking up real data from SiteDataIndex; after being resolved, the SiteDataReference will always identify the same data, even if its referenceable key changes. * (That in turn required rearchitecture of SiteDataIndex internals; data dictionaries are no longer keyed by mutable ReferenceableKeys, but by immutable DataIds instead.)
Current problem: entities with the LinkedPost concept, which is a list of PostData, attempt to render those posts' metadata (categories and games and platforms) in string templates; but those metadata items are no longer raw strings, so the rendered output is now csharp typenames.  Need to refactor the LinkedPosts concept so these links can consist of pure metadata, instead.

Another todo: the SiteDataReferences pulled out of PostData FrontMatter aren't yet *put back into* the FrontMatter; this needs to happen for the Toml FrontMatter part of a post to roundtrip-render post-modification reference keys.

And then: GlogLinks (from markdown parsing) should get a SiteDataReference instance, so that the referenceable key / permalink can be looked up when the GlogLink is rendered, including during roundtrip rendering.  (To be unit tested by modifying the referenced data's key after parsing it.)