Some work-in-progress code I've been using to auto-detect and index "related games" from IGDB (bundles, remasters, DLCs, the whole lot).
This is where all that auto-disambiguation I've been hacking on will come in -- so I don't have to manually distinguish each and every re-release of DS games remastered for mobile and remastered again for PC ... et al.

But, as it stands this adds a LOT of data to the index and a lot of empty pages to the site (since I never wrote a post about most of these games).  I want to do some index-culling pass so that un-referenced games/etc don't get persisted in the index and don't result in empty Glog pages.

And I should probably do that culling change first; if the new auto-collected data comes first, then the diff from there to the culled output will be way too big to make sense of.