Yesterday's tag-update fail was really due to two separate problems: 1. TagData isn't clearly correlated to original IgdbEntity IDs, and hence tag references aren't so straightforward to update when an IGDB tag string is modified.  ... but this has always been the case, and is difficult to un-tangle since TagData may map to *multiple* IgdbEntities, which may not even be updating consistently with one another.  (Should probably still fix that, but that's, eugh.) 2. Recent unreferenced-data-culling work had to include additional "throwaway" CreateReference() calls (i.e. new SiteDataIndex DataReference registrations) in order to avoid culling a Game's tags, even when those tags aren't explicitly referenced in content (such as posts).  This behavior of registering not-explicitly-referenced tag strings is new!
That second behavior, of treating all game-associated tags just like content-referenced tags, wasn't quite appropriate -- because when IGDB data is updated and SiteDataIndex is re-built, old tags which weren't explicitly referenced by name SHOULD be discarded! because attempting to update their reference strings wouldn't serve any purpose anyway.

So, now it's possible to add a DataReference to the SiteDataIndex which *will not* be updated in LoadContent(), and will instead be discarded (possibly replaced with new data, or possibly gone forever).  These should-not-update-on-data-change DataReferences only exist for the purpose of marking data as in-use, to avoid culling in RemoveUnreferencedData, so that they'll be used when generating site content / rendering post-list pages.