Goal: new build mode `UpdateDataNoOutput` and new option `add-igdb-game-ids` can add games (by ID) to the IGDB cache before/during its data update.
Why?: because since adding metadata-dependent title (and inline reference) behavior, now adding a game isn't necessarily as simple as pasting an ID and name into the .json cache file; if the game is a re-release, for example, its cache entry needs IGDB's release date before the true name can be resolved.

So with this change, it should be possible to add the Castlevania Advance Collection's 2021 iteration of "Castlevania: Harmony of Dissonance" by just adding game ID 222411 while building the data index.

This isn't really working in practice yet, though, because a reference to e.g. "Castlevania: Harmony of Dissonance (2021)" is invalid before the data cache has it -- so when a reference exists before the data is updated, the site build fails.  And on the flip side, if a reference doesn't exist, unreferenced-data culling will remove that game ID immediately after a data update adds it.  ... hmmmmmm.