Massaging the new templates to FINALLY match* existing output.
(*) Except now there's always a trailing newline at the end of the HTML file, which is definitely more correct than not.  (I didn't know this was missing!)  Hence the TestFiles expected-output update.

Some of the line-break changes here seem very silly, so there is still a little readability work left.