A starter implementation of SiteDataIndex state persistence (to replace the committed IGGB cache).
After so, so many IGDB cache updates that didn't need to be version controlled - pull requests with huge JSON diffs churning virtually meaningless metadata - I finally realized that those cache files have been conflating multiple purposes:

1. Some entity properties are needed to build the SiteDataIndex, and help it understand how content refers to data entities. These properties are a combination of IGDB metadata and occasional Glog override settings.

For example: when a post includes a hyperlink to Some Game, these properties are used to associate the string "Some Game" to the game's IGDB entity and that entity's metadata; this allows GlogGenerator to detect when that IGDB entity's name is changed to e.g. "s0me gam3" and the post's hyperlink can be rewritten with that updated reference string.

2. Other entity properties are needed to fully hydrate the site structure, and render HTML pages for the website build. These properties are sourced solely from IGDB.

For example: posts for Some Game will be listed on a page that shows additional detail about Some Game, including its associated franchise and genre, remasters, bundles, DLC... as extracted from IGDB's database. Pages will also be generated for that franchise, for that genre, for each remaster, *et cetera*.

The IGDB cache files currently in this repository handle both purposes. This allows repo files as-committed to build the SiteDataIndex *and* build the website without requiring a fresh pull of data from IGDB's APIs.

In practice though, when GitHub Actions is building the website, it's always pulling fresh IGDB data anyway.

And while there's still a development use-case for the full IGDB cache (when I'm building a preview or debug version of the site at home), I can generate that cache ad-hoc; the repo's copy isn't needed.

*Except* for the 1st purpose's entity properties, which are still needed to preserve referential integrity as IGDB names change. In other words: point 2's full IGDB cache doesn't need to be committed, just the subset of metadata which can fulfill point 1.

GlogGenerator just needs to be able to build a SiteDataIndex from that subset of metadata -- then during an update of IGDB metadata, GlogGenerator can revise and rewrite that minimal set of data when needed.

I'm starting with the "rewrite" part. This change implements SiteDataIndex serialization to record a small set of metadata for each reference-able entity (for Games, Platforms, and generic Tags), metadata which is sufficient for reconstructing the string which content uses to reference that entity.

Later changes will implement the ability to build GlogGenerator's SiteDataIndex *from* that recorded metadata, without a full IGDB cache; and then ultimately the cache files can be removed from the glog's saved state without consequence.