Juggling non-IGDB entity data around for clarity.
In the case of 6 games and 1 platform which don't really exist in IGDB, but I've invented data for them in the IGDB cache ... going forward, their not-from-IGDB state will be clearer in SiteDataIndex storage. Meanwhile, indicating their fake data properties as "overrides" (on top of, well, nothing) is how that SiteDataIndex storage will get the message.