Fixing the AddGames verb.
Two abstract changes were required for this fix:

1. The verb itself now has a smaller scope; it no longer attempts to fully "load" data, parse content, and glom the added game(s) onto a routine IGDB API update. Now it much more surgically...
   a. Deserializes the SiteDataIndex: `ISiteDataIndex.ReadJsonFiles()`
   b. Executes a very small IGDB API fetch for just the requested game ID(s).
   c. Adds the relevant entity reference properties to the index: `ISiteDataIndex.AddGameReferences()`
   d. And re-serializes the SiteDataIndex before exiting: `ISiteDataIndex.WriteJsonFiles()`

2. I missed something in the recent work to separate SiteDataIndex persistence from IgdbCache persistence: the case where a cache *exists* but is out-of-date. Was bound to run into this gap sooner or later!

And the new AddGames implementation highlighted the problem pretty clearly: after adding an IGDB entity ID to `sitedataindex/*`, the next site build - using whatever's left over in `igdbcache/*` - won't be able to validate that entity reference. The cache needs to be updated, first.

This is fixed by replacing the program's "if the cache was loaded, an update isn't necessary" assumption with an explicit `SiteBuilder.DataIndexRequiresIgdbCacheUpdate()` boolean check.